% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_shoemaker2015}
\alias{load_shoemaker2015}
\title{Loading the shoemaker2015 dataset as SummarizedExperiment}
\usage{
load_shoemaker2015()
}
\description{
See data(shoemaker2015) for a full description of the dataset and the
metadata.
}
\examples{
mice_timecourse_data = load_shoemaker2015()
head(mice_timecourse_data)
}
