% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_varoquaux2019}
\alias{load_varoquaux2019}
\title{Loading the varoquaux2019 leaf or root dataset as SummarizedExperiment}
\usage{
load_varoquaux2019(sample_type = "leaf")
}
\arguments{
\item{sample_type}{{"leaf", "root"}, optional, default: leaf
Sets which sample to load.}
}
\value{
a \code{SummerizedExperiment} of the leaf or root samples of
Varoquaux et al.
}
\description{
See data(varoquaux2019leaf) or data(varoquaux2019root) for a full
description of the dataset and the metadata.
}
\examples{
sorghum_timecourse_data = load_varoquaux2019(sample_type="leaf")
head(sorghum_timecourse_data)
}
