% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_varoquaux2019leaf}
\alias{load_varoquaux2019leaf}
\title{Loading the varoquaux2019leaf dataset as SummarizedExperiment}
\usage{
load_varoquaux2019leaf()
}
\description{
See data(varoquaux2019leaf) for a full description of the dataset and the
metadata.
}
\examples{
sorghum_timecourse_data = load_varoquaux2019leaf()
head(sorghum_timecourse_data)
}
