% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_varoquaux2019root}
\alias{load_varoquaux2019root}
\title{Loading the varoquaux2019root dataset as SummarizedExperiment}
\usage{
load_varoquaux2019root()
}
\value{
à \code{SummerizedExperiment} of the root samples of Varoquaux et
al.
}
\description{
See data(varoquaux2019root) for a full description of the dataset and the
metadata.
}
\examples{
sorghum_timecourse_data = load_varoquaux2019root()
head(sorghum_timecourse_data)
}
