% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varoquaux2019root.R
\docType{data}
\name{varoquaux2019root}
\alias{varoquaux2019root}
\title{Time course transcriptomic data from root tissues exposed to drought stress}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(varoquaux2019root)
}
\description{
Time-course RNA-seq root data from "Transcriptomic analysis of field-droughted
sorghum from seedling to maturity reveals biotic and metabolic responses"
}
\examples{
data(varoquaux2019root)
data = attr(varoquaux2019root, "data")
meta = attr(varoquaux2019root, "meta")
}
\references{
Varoquaux et al, 2019
(\href{https://www.pnas.org/content/116/52/27124.short}{PNAS})
}
\keyword{datasets}
