% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varoquaux2019leaf.R
\docType{data}
\name{varoquaux2019leaf}
\alias{varoquaux2019leaf}
\title{Time course transcriptomic data from leaf tissues exposed to drought stress}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(varoquaux2019leaf)
}
\description{
Time-course RNA-seq leaf data from "Transcriptomic analysis of field-droughted
sorghum from seedling to maturity reveals biotic and metabolic responses"
}
\examples{
data(varoquaux2019leaf)
data = attr(varoquaux2019leaf, "data")
meta = attr(varoquaux2019leaf, "meta")
}
\references{
Varoquaux et al, 2019
(\href{https://www.pnas.org/content/116/52/27124.short}{PNAS})
}
\keyword{datasets}
